Added MiniMax.kt and updated GameScreen.kt
The minimax.kt has 3 functions, runAITurn which is a controller and does the easy difficulty return. It then calls the miniMax function for the medium and hard difficulties, which runs the minimax algorithm with alpha-beta pruning. The final function is needed to figure out when someone wins so that the miniMax algorithm is able to know how to assign values. The depth is there to cause earlier wins to be prioritized over later ones. The game code still needs to be updated to display a win and pass the difficulty to the controller.