AITurn with X or O player
Updated minimax and AITurn to work with X or O player by passing them the value of the players choice at the start of a game or the setting they choose. The functionality is the same the CPU just is now able to check for what it is supposed to be placing and what it is supposed to be maximizing.